+++ plugins/gs/src/apis/types.ts
@@ -1,19 +1,3 @@
-type BaseRequestResult = {
-  installationName: string;
-}
-
-export type FulfilledRequestResult<T> = BaseRequestResult & {
-  status: 'fulfilled',
-  value: T[];
-};
-
-export type RejectedRequestResult = BaseRequestResult & {
-  status: 'rejected',
-  reason: any;
-}
-
-export type RequestResult<T> = FulfilledRequestResult<T> | RejectedRequestResult;
-
 export type Resource<T> = T & {
   installationName: string;
 }