+++ plugins/gs/src/components/InstallationsWrapper/index.ts
@@ -1,0 +1,1 @@
+export { InstallationsWrapper } from './InstallationsWrapper';